rewrite structs as wrappers
